Replace custom publish script with lerna publish
The `--force-publish=*` option is specified to ensure that all
packages are released, rather than just those that were modified, to
preserve the Neutrino tradition of keeping all versions in sync.

The new `release:preview` command is unfortunately not a true dry run,
in that it leaves changes in the working directory that have to be
reverted, but at least allows for confirming the necessary versions
have been adjusted.

With the `.scripts/publish` script removed, the `graph-data-structure`
and `semver` devDependencies are now unused.