Follow-ups to the switch to Lerna (#372)
* Remove redundant `lerna bootstrap` call on Travis

Due to yarn workspaces, the main yarn install command (already run
prior to bootstrap on Travis) itself handles the linking/bootstrapping
of dependencies. As such, `lerna bootstrap` is now just an alias for
`yarn install`, and therefore redundant.

* Remove yarn.lock from the package's .npmignore files

Since with yarn workspaces, the per-package yarn.lock file no longer
exists, with all dependencies instead pinned from the root yarn.lock.

* Set node and yarn engine versions in package.json

To make it more obvious to users/contributors when using versions
that have not been tested (or in the case of yarn, that don't support
the workspaces feature).

* Update development docs with lerna/workspaces changes

Since we're no longer using oao. The reference to creating package
links has been removed since yarn workspaces intentionally constrains
development versions to the individual workspace, rather than
allowing them to affect other projects installed elsewhere.

* Replace custom publish script with lerna publish

The `--force-publish=*` option is specified to ensure that all
packages are released, rather than just those that were modified, to
preserve the Neutrino tradition of keeping all versions in sync.

The new `release:preview` command is unfortunately not a true dry run,
in that it leaves changes in the working directory that have to be
reverted, but at least allows for confirming the necessary versions
have been adjusted.

With the `.scripts/publish` script removed, the `graph-data-structure`
and `semver` devDependencies are now unused.

* Remove version from root package.json

Since the global version is now tracked via lerna.json, and lerna
doesn't know to update the root package.json, so it will only end up
getting out of sync.

* Remove unnecessary devDependencies from root package.json

Now that we're using yarn workspaces, linting of the repo is possible
without having to duplicate the `neutrino-preset-airbnb-base` and
`neutrino-middleware-eslint` sub-dependencies into the repo root
`package.json`, since they are all available in the root `node_modules`.

* Regenerate yarn.lock