Remove unnecessary monorepo linting rule overrides
These rules are now either set via the AirBnb or prettier presets,
or else are no longer required. Fewer overrides makes it easier
to reason about the actual differences between AirBnB/prettier and
the linting config used for this monorepo.

The only code fix required was making `_logo()` be a static method
in the create-project init command, to fix:
https://eslint.org/docs/rules/class-methods-use-this