Cleanup monorepo linting configuration (#758)
* Remove React presets/rules from monorepo lint config

Since we don't currently lint any React files, and if we start doing so
in the future we should use the airbnb preset instead of airbnb-base.

* Remove unnecessary monorepo linting rule overrides

These rules are now either set via the AirBnb or prettier presets,
or else are no longer required. Fewer overrides makes it easier
to reason about the actual differences between AirBnB/prettier and
the linting config used for this monorepo.

The only code fix required was making `_logo()` be a static method
in the create-project init command, to fix:
https://eslint.org/docs/rules/class-methods-use-this

* Make monorepo linting file includes less fragile

Switches from opt-in to opt-out to avoid accidentally omitting new
directories or non-JS file types. The list of files scanned is the
same before/after (confirmed using `yarn lint --debug`).

* Add eslint to monorepo root devDependencies

This prevents warnings during yarn install of the monorepo:
```
warning " > eslint-config-prettier@2.9.0" has unmet peer dependency "eslint@>=3.14.1".
```