[v8.x] Unpin the fast-async version
The `fast-async` version was pinned in 2edc81a to work around:
https://github.com/MatAtBread/fast-async/issues/53

That issue was fixed in fast-async v6.3.4, so it's safe to revert
back to a tilde range again.

This only needs applying to the v8 branch, since the version was not
pinned on master (and #790 is about to remove fast-async there).