Replace @neutrinojs/chunk with splitChunks
webpack 4 now has an intelligent chunking feature that replaces
`CommonsChunkPlugin` and avoids the need to create the fake `vendor`
entrypoint to separate out third-party dependencies:
https://webpack.js.org/plugins/split-chunks-plugin/

In addition, several of the plugins previously used are automatically
enabled when `mode` is development:
https://github.com/webpack/webpack/blob/v4.6.0/lib/WebpackOptionsDefaulter.js#L250-L259
https://github.com/webpack/webpack/blob/v4.6.0/lib/WebpackOptionsApply.js#L325-L328
(Note this differs slightly from our previous implementation, but if
this results in long-term caching issues, issues should be filed to
change the defaults upstream, and only set here as a last resort).

However since `splitChunks` creates an unknown number of chunks with
unpredictable filenames, it's no longer possible to pass an exact
chunks list to `html-webpack-plugin`'s `chunks` option. We also can't
omit the `chunks` option entirely, since for multi-entrypoint builds
that would result in all assets being included for every page.

As such, until `html-webpack-plugin` natively supports `splitChunks`,
we have to use `html-webpack-include-sibling-chunks` to dynamically
generate the `chunks` list:
https://github.com/fenivana/html-webpack-include-sibling-chunks-plugin

With the switch to `splitChunks`, `@neutrinojs/chunk` becomes simple
enough that it's not really worth keeping separate from the web preset,
particularly given that it relies on the sibling chunks workaround to
actually be useable anyway.

Note that webpack enables `splitChunks` in development too (which is
great for dev-prod parity, and fixes a bug I was hitting migrating
an app to Neutrino 8), which is why our overrides are no longer
inside the `NODE_ENV === 'production'` block.