Use uglify-es instead of babel-minify by default
`uglify-js` (used by webpack 3) doesn't support ES6, so previously the
web preset used `babel-minify` instead, to avoid having to transpile
the generated javascript down to ES5 before minification.

However webpack 4 now uses the new `uglify-es`, which does support ES6,
and is much faster than `babel-minify`. As such, the web preset now
defaults to `uglify-es` (which webpack 4 enables by default when `mode`
is production).

In addition, `babel-minify-webpack-plugin` has been found to be buggy
when combined with `webpack-sources` > `1.0.1`, which required hacky
workarounds in the form of yarn `resolutions` overrides and a forked
version of `babel-minify-webpack-plugin`. This means we're on an old
version of `webpack-sources`, which is missing fixes and performance
improvements.

It's likely before this PR is merged (or else before Neutrino 9 is
released) that `@neutrinojs/babel-minify` will be completely detached
from the web preset (and the workarounds removed) - but for now I've
left it in, for easier A/B testing of the two minifiers.

For more context, see:
https://github.com/mozilla-neutrino/neutrino-dev/issues/748

It's also looking like we may want to adjust `webpack-chain` in the
future, so that `optimization.minimizer` behaves like `plugins` and
can have named entries for easier overriding. (To allow setting both
JS and CSS minifiers there at once - especially given webpack 5 may
also include a CSS minifier there too.)