Docs: Update example .{es,style}lintrc.js shims to set explicit root directory (#703)
When emacs flycheck package runs `eslint`, it runs it from the same
directory as the source file.  If that's the case, the suggested
`.eslintrc.js` shim in this file will fail to find the `.neutrinorc.js`
file.  This simple fix just looks for it in the same directory as the
`.eslintrc.js` file.  This shouldn't (hopefully) have any negative
consequences for any IDEs that run `eslint` from the project root that
are currently working.