Remove @neutrinojs/babel-minify
We're going to stick with webpack 4's default of uglify-es for now,
since it's much faster than babel-minify. Between that and the new
webpack `optimization.minimize` and `optimization.minimizer` options,
it's simpler to just document how to override the defaults directly,
rather than provide a separate preset per alternative minifier.