preact: Enable eslint-plugin-react if using ESLint
`@neutrinojs/preact` already had a dependency on `eslint-plugin-react`,
but wasn't actually using it. Now the React ESLint plugin is enabled
if an ESLint configuration is found - the same as for the React preset.

This will make it easier for Preact users to realise what changes they
need to make, now that we're not using `babel-plugin-jsx-pragmatic`.

Fixes #821.