Remove legacy gitbook functionality
Now that we're using the new hosted version of Gitbook, there is no
need to build the UI locally (and they don't provide a tool to be
able to do so anyway) - so we don't need any of the gitbook-cli
related dependencies or gh-pages for publishing.

I've removed the old `book.json` configuration file, and replaced
it with the new `.gitbook.yaml`, in case they were silently using
the former to determine our repository layout. See:
https://docs.gitbook.com/integrations/github#content-configuration