[v8.x] Backport stringified config improvements as --inspect-new
This backports #928 to `release/v8` under a new argument name, to
prevent it from being a breaking change. This will allow Neutrino 8
users to benefit from the stringifed config improvements sooner, as
well as providing a way to compare v8 and v9 configurations.

I've not backported the eslint `formatter.__expression` change, since
it would require a fair amount of refactoring, increasing risk for
little benefit.

Fixes #328 for `release/v8`.