eslint: Throw if preset used after compile rule already defined
Presets such as `@neutrinojs/web` and `@neutrinojs/react` check to
see if a `lint` rule is defined, and if so, adapt its settings to
suit the particular environment/framework. This means that the lint
presets must be defined prior to the compile presets or the final
eslint configuration will not be correct.

However previously this broken ordering would not result in an error
message, causing confusion when eslint wasn't behaving as expected.

Fixes #800.
Refs #763.