[v8.x] Add object rest spread support to web & library presets
Previously the `babel-plugin-transform-object-rest-spread` plugin was
only enabled for `@neutrinojs/react` and `@neutrinojs/preact` and not
`@neutrinojs/web` or any of the other presets that inherit from it.

This inconsistency is both confusing for users, and means having to
manually add the plugin when ideally support for object rest spread
syntax would be enabled out of the box - now that it's stage 4
(when Neutrino 8 was released, it was stage 3):
https://github.com/tc39/proposal-object-rest-spread#status-of-this-proposal

Duplicate plugin instances didn't cause errors when I tested locally,
so this should be backwards compatible for users/presets that already
manually add the plugin.

This change isn't needed on `master`, since Babel 7's `preset-env`
ships with `proposal-object-rest-spread` built in.

Refs #883.