html-loader: Add 'link:href' to attrs so favicons are parsed
By default `html-loader@0.x` has a default `attrs` of `['img:src']`,
which means only `<img src="...">` tags are followed when parsing
the HTML:
https://github.com/webpack-contrib/html-loader/blob/v0.5.5/README.md#usage
https://github.com/webpack-contrib/html-loader/blob/v0.5.5/index.js#L29

With this change, favicons are included too, and have all the benefits
of being included in the webpack build - such as being being inlined
by url-loader (if small enough), or else being copied to the output
directory automatically and given hashed filenames.

This matches the default that will be used in `html-loader` v1.0.0:
https://github.com/webpack-contrib/html-loader/issues/17

...and saves projects from having to resort to the manual Neutrino
API to adjust the options used by `html-loader` (see #865).