Testing: Stop using lerna canary mode for publishing
Since we need the `package.json` version bumps to be left in the
working directory (which doesn't happen with canary mode), so that
`create-project` can read the correct version when determining the
matching monorepo package versions.

This also works around the bug where Verdaccio rejects alphanumeric
prerelease identifiers that start with a number (ie a proportion of
SHAs).

The `--npm-tag latest` is no longer required, since non-canary mode
defaults to `latest`.