Use lerna exact mode & bump version to 9.0.0-0 prerelease (#958)
* Clean up lerna configuration

- Moves `--force-publish` to the config file.
- Removes redundant `packages` option (since `useWorkspaces` set).
- Moves `gitRemote` under the publish command options, since it's
  not used by any other commands.

* Use lerna exact mode when publishing

This makes lerna use exact versions rather than caret ranges when
updating the inter-monorepo package dependency versions, which makes
it possible to install historic versions of monorepo packages without
them using a mixture of old and new minor/patch versions.

See:
https://github.com/lerna/lerna#--exact-1

* Bump version to 9.0.0-0 prerelease

To make it clearer that `master` is for the upcoming Neutrino 9, and
also start using exact versions. The peer dependencies had to be
updated manually, and have to use a pre-release identifier to prevent
yarn warnings spam when using presets with pre-release `neutrino`.