Add `transform-object-rest-spread` plugin for Node
Since the Node preset already includes support for most modern Ecmascript features including async functions, `import`, etc., the lack of support for the spread operator is confusing and seems out of place.  This commit adds support for the spread operator by default, allowing for it to be parsed properly when users use it in their source code.