create-project: Update webpack-cli from v2 to v3
We hadn't received a Renovate PR for this `webpack-cli` release (which
would have reminded us to bump the version in `matrix.js`), since the
dependency is intentionally not listed in `peerDependencies` so that
people can install their own choice of CLI if desired. To prevent this
problem in the future, `webpack-cli` has now been added to the monorepo
`devDependencies`, so we'll get the reminder PR next time.

https://github.com/webpack/webpack-cli/compare/v2.1.5...v.3.1.0