jest: Use babel-jest instead of a custom implementation
We were previously just reimplementing parts of it, and not using
the latest recommended Babel settings (such as still setting
`retainLines: true` when sourcemaps are supported natively).

There isn't currently a way to set the babel-jest `createTransformer()`
options via Jest configuration/globals, so we instead have to resort
to using an environment variable. This has the limitation of only
supporting Babel options that are serializable (virtually all commonly
used configuration options are), with any others being silently omitted.
However this was already the case when using `globals` anyway.

See:
https://github.com/facebook/jest/blob/v23.6.0/packages/babel-jest/README.md
https://github.com/facebook/jest/blob/v23.6.0/packages/babel-jest/src/index.js

Fixes #851.