compile-loader: Remove babel-merge export
Since it's no longer a custom implementation and instead just the
`babel-merge` package exported. As such it makes more sense for
packages to import `babel-merge` directly iff they need it.

This also avoids the naming clash between `merge` from this preset
and the typical import style used for `deepmerge` in this repository,
that contributed to the bug fixed by #1219.