create-project: Run correct command for initial lint when using NPM
Previously the `--fix` option was not passed to ESLint when using NPM,
since it requires that any options passed be preceded with `--`.

This caused failures during create-project when users picked a lint preset
whose style differed from the template project, since the failures would
not be fixed, and instead ESLint exit non-zero causing create-project to
abort and delete the new project.

Fixes #1261.