Update eslint-config-standard and switch from -jsx to -react
Updates the ESLint standardjs presets to latest, and switches from
`eslint-config-standard-jsx` to `eslint-config-standard-react`, since
the latter is a superset of the former and required for full React
support.

The new configs set the React version and plugins for us, so we no
longer need to set them ourselves.

Fixes #1449 and fixes #1450.