react: Use babel-preset-react's new useSpread option
A new `useSpread` option was added in Babel 7.7.0 (babel/babel#10572):
https://babeljs.io/docs/en/babel-preset-react#usespread

Enabling this means that the JSX-> JS conversion will use an inline
object with spread elements (for code that spreads props), rather than
Babel's extend helper or `Object.assign`.

`@babel/env` will still transpile these down if required for the target
browsers (the only reason `useSpread` is not already enabled by default
is that for the upstream project this is a breaking change, since they
cannot assume that users of `babel-preset-react` are always using it
alongside `@babel/env`, even though most likely are).